Moved ContentPaginationState to common_activity_states.py module as it has nothing specific to do with message_handler_voice
